requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.6
djangorestframework==3.15.1
idna==3.7
pymessenger==0.0.7.0
python-dotenv==1.0.0
requests==2.32.3
requests-toolbelt==1.0.0
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.2.1
